third_party/WebKit/LayoutTests/fast/url/encoding-euckr.html

<!DOCTYPE html>
<meta charset="euc-kr">
<title>Verify that NFC normalization is not done - EUC-KR</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
test(function(t) {
    assert_equals(document.charset, 'EUC-KR');

    var a = document.createElement('a');
    a.href = 'https://example.com/?x\u6A02\uF914x';
    assert_equals(a.search.substr(1), 'x%E4%C5%D1%E2x');
}, 'NFC normalization is not done prior to encoding');
</script>
